backend/requirements.txt.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
django-cors-headers==4.3.1
Pillow==10.1.0